requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic-settings==2.6.1
python-dotenv==1.0.1

pyarabic==0.6.15
arabic-reshaper==3.0.0

sentence-transformers==3.3.1
chromadb==0.5.23
rank-bm25==0.2.2

google-generativeai==0.8.3

pypdf==5.1.0
beautifulsoup4==4.12.3
requests==2.32.3

ragas==0.2.6
datasets==3.2.0

pytest==8.3.4
httpx==0.28.1